refactor: rename token select step file and update import
